chefusion-mini-claude/fusion_cuisine_poc/pyproject.toml
[tool.poetry]
name = "fusion-cuisine-poc"
version = "0.1.0"
description = "Multicultural Fusion Cuisine Generator & Palate Predictor"
authors = ["Claude AI <claude@anthropic.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.3.0"
torchvision = "^0.18.0"
transformers = "^4.30.0"
diffusers = "^0.18.0"
datasets = "^2.14.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
pillow = "^10.0.0"
requests = "^2.31.0"
tqdm = "^4.65.0"
pyyaml = "^6.0.0"
openai = "^1.0.0"
accelerate = "^0.20.0"
scipy = "^1.11.0"
sentence-transformers = "^2.2.0"
jupyterlab = "^4.0.0"
jupyter = "^1.0.0"
ipywidgets = "^8.0.0"
plotly = "^5.15.0"
nltk = "^3.8.0"
spacy = "^3.7.0"
networkx = "^3.1.0"
pytorch-lightning = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true